Added knex-store
Need some changes
